requirements.txt
langchain>=0.2.15
langchain-community>=0.2.15
langchain-core>=0.2.35
langchain-text-splitters>=0.2.2
langchain-ollama>=0.1.0
chromadb>=0.5.5
pypdf>=4.2.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
streamlit>=1.36.0

